rearange includes and forward decls in ParseFrame.h and uncrustify_types.h
making frames a vector allows `ParseFrame` to be forward declared in
uncrustify_types.h
-> allows inclusion of 'uncrustify_types.h' inside 'ParseFrame.h' without
   inclusion loops
  -> makes it possible to move `brace_stage_e` back into 'uncrustify_types.h',
     and `paren_stack_entry_t` from 'uncrustify_types.h' into 'ParseFrame.h'
    -> shared_ptr hack for `last_poped` can now be removed
